fix: items 컴포넌트 분리 작성
