+++ src/components/items/AllProductsHeader.module.css
@@ -1,34 +1,7 @@
 .container {
+  width: 1200px;
   margin: 0 auto;
   margin-top: 40px;
-  width: 1200px;
-}
-
-.all-product-list {
-  display: flex;
-  flex-wrap: wrap;
-  row-gap: 40px;
-  column-gap: 24px;
-}
-
-.all-product {
-  width: 220px;
-  height: 317px;
-}
-
-.all-product-info {
-  width: 217px;
-  display: flex;
-  flex-direction: column;
-  row-gap: 6px;
-}
-
-.all-product-img {
-  width: 220px;
-  height: 220px;
-  border-radius: 16px;
-  padding-bottom: 16px;
-  cursor: pointer;
 }
 
 .title-container {
@@ -119,102 +92,14 @@
   background-color: var(--secondary-300);
 }
 
-.product-title {
-  color: var(--secondary-800);
-  font-size: 14px;
-  font-style: normal;
-  font-weight: 500;
-  line-height: 24px;
-  cursor: pointer;
-}
-
-.product-price {
-  color: var(--Secondary-800);
-  font-size: 16px;
-  font-style: normal;
-  font-weight: 700;
-  line-height: 26px;
-}
-
-.product-favorite-count {
-  color: var(--Secondary-600);
-  font-size: 12px;
-  font-style: normal;
-  font-weight: 500;
-  line-height: 18px;
-  display: flex;
-  gap: 4px;
-}
-
-.product-heart-img>svg {
-  width: 16px;
-  height: 16px;
-  cursor: pointer;
-}
-
-.pagination {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  gap: 4px;
-  margin-top: 43px;
-  margin-bottom: 58px;
-}
-
-.pagination span {
-  width: 40px;
-  height: 40px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  border: 1px solid var(--Cool-Gray-200, #E5E7EB);
-  ;
-  border-radius: 40px;
-  font-size: 15px;
-  cursor: pointer;
-  user-select: none;
-}
-
-.pagination>span:hover {
-  transition: background-color 0.3s ease-out;
-  background-color: var(--secondary-300);
-}
-
 @media (max-width: 1199px) {
   .container {
     width: 100%;
     padding: 0 24px;
   }
 
-  .all-product-list {
-    display: flex;
-    flex-wrap: wrap;
-    column-gap: 16px;
-  }
-
-  .all-product {
-    width: 100%;
-    flex: 1 1 30%;
-  }
-
-  .all-product-info {
-    width: 100%;
-    flex: 1 1 30%;
-  }
-
-  .all-product-img {
-    width: 100%;
-    flex: 1 1 30%;
-    border-radius: 12.9px;
-  }
-
   .search-input {
     width: 262px;
-  }
-
-  .pagination {
-    margin-top: 40px;
-    margin-bottom: 72px;
   }
 }
 
@@ -223,30 +108,7 @@
     width: 100%;
     padding: 0 16px;
     max-width: 400px;
-  }
-
-  .all-product-list {
-    margin-top: 71px;
-    display: flex;
-    column-gap: 8px;
-  }
-
-  .all-product {
-    width: 100%;
-    flex: 1 1 45%;
-    height: 264px;
-  }
-
-  .all-product-info {
-    width: 100%;
-    flex: 1 1 45%;
-  }
-
-  .all-product-img {
-    width: 100%;
-    flex: 1 1 45%;
-    height: 167px;
-    border-radius: 9.82px;
+    margin-top: 24px;
   }
 
   .title-container {
@@ -255,7 +117,7 @@
 
   .search-container {
     position: absolute;
-    top:0;
+    top: 0;
     display: flex;
     justify-content: flex-end;
     flex-wrap: wrap;
@@ -304,10 +166,5 @@
   .search-dropdown-list {
     top: 36px;
     right: 0;
-  }
-
-  .pagination {
-    margin-top: 40px;
-    margin-bottom: 35px;
   }
 }
