+++ src/components/items/BestProducts.module.css
@@ -7,87 +7,10 @@
   margin-bottom: 16px;
 }
 
-.best-product-list {
-  display: flex;
-  gap: 24px;
-}
-.best-product {
-  display: flex;
-  flex-direction: column;
-  gap: 16px;
-  width: 282px;
-  height: 379px;
-}
-
-.best-product-info {
-  display: flex;
-  flex-direction: column;
-  gap: 6px;
-}
-
-.best-product-img {
-  width: 282px;
-  height: 282px;
-  border-radius: 16px;
-  cursor: pointer;
-}
-
-.product-heart-img > svg {
-  width: 16px;
-  height: 16px;
-  cursor: pointer;
-}
-
-.product-title {
-  color: var(--secondary-800);
-  font-size: 14px;
-  font-style: normal;
-  font-weight: 500;
-  line-height: 24px;
-  cursor: pointer;
-}
-
-.product-price {
-  color: var(--Secondary-800);
-  font-size: 16px;
-  font-style: normal;
-  font-weight: 700;
-  line-height: 26px;
-}
-
-.product-favorite-count {
-  color: var(--Secondary-600);
-  font-size: 12px;
-  font-style: normal;
-  font-weight: 500;
-  line-height: 18px;
-  display: flex;
-  gap: 4px;
-}
-
 @media (max-width: 1199px) {
   .container {
     width: 100%;
     padding: 0 24px;
-  }
-  
-  .best-product-list {
-    display: flex;
-    justify-content: space-between;
-  }
-
-  .best-product {
-    display: flex;
-    flex-direction: column;
-    gap: 10px;
-    width: 100%;
-    height: 434px;
-  }
-  
-  .best-product-img {
-    width: 100%;
-    height: 343px;
-    border-radius: 20.2px;
   }
 }
 
@@ -96,43 +19,5 @@
     width: 100%;
     padding: 0 16px;
     max-width: 400px;
-  }
-  
-  .best-product {
-    display: flex;
-    flex-direction: column;
-    gap: 10px;
-    width: 100%;
-    height: 434px;
-  }
-  
-  .best-product-img {
-    width: 100%;
-    height: 343px;
-    border-radius: 20.2px;
-  }
-  
-  .product-title {
-    font-size: 14px;
-    font-style: normal;
-    font-weight: 500;
-    line-height: 24px;
-    cursor: pointer;
-  }
-  
-  .product-price {
-    font-size: 16px;
-    font-style: normal;
-    font-weight: 700;
-    line-height: 26px;
-  }
-  
-  .product-favorite-count {
-    font-size: 12px;
-    font-style: normal;
-    font-weight: 500;
-    line-height: 18px;
-    display: flex;
-    gap: 4px;
   }
 }
