+++ src/components/items/ItemPageNation.module.css
@@ -1,0 +1,41 @@
+.pagination {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 4px;
+  margin-top: 43px;
+  margin-bottom: 58px;
+}
+
+.pagination span {
+  width: 40px;
+  height: 40px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  border: 1px solid var(--Cool-Gray-200, #E5E7EB);
+  ;
+  border-radius: 40px;
+  font-size: 15px;
+  cursor: pointer;
+  user-select: none;
+}
+
+.pagination > span:hover {
+  transition: background-color 0.3s ease-out;
+  background-color: var(--secondary-300);
+}
+
+@media (max-width: 1199px) {
+  .pagination {
+    margin-top: 40px;
+    margin-bottom: 72px;
+  }
+}
+
+@media (max-width: 767px) {
+  .pagination {
+    margin-top: 40px;
+    margin-bottom: 35px;
+  }
+}
