+++ src/components/items/ProductCard.module.css
@@ -1,0 +1,184 @@
+.container {
+  width: 1200px;
+  margin: 0 auto;
+}
+
+.best-product-list {
+  display: flex;
+  gap: 24px;
+}
+
+.best-product {
+  display: flex;
+  flex-direction: column;
+  gap: 16px;
+  width: 282px;
+  height: 379px;
+}
+
+.best-product-info {
+  display: flex;
+  flex-direction: column;
+  gap: 6px;
+}
+
+.best-product-img {
+  width: 282px;
+  height: 282px;
+  border-radius: 16px;
+  cursor: pointer;
+}
+
+.all-product-list {
+  display: flex;
+  flex-wrap: wrap;
+  row-gap: 40px;
+  column-gap: 24px;
+}
+
+.all-product {
+  width: 220px;
+  height: 317px;
+}
+
+.all-product-info {
+  width: 217px;
+  display: flex;
+  flex-direction: column;
+  row-gap: 6px;
+}
+
+.all-product-img {
+  width: 220px;
+  height: 220px;
+  border-radius: 16px;
+  padding-bottom: 16px;
+  cursor: pointer;
+}
+
+.product-title {
+  color: var(--secondary-800);
+  font-size: 14px;
+  font-style: normal;
+  font-weight: 500;
+  line-height: 24px;
+  cursor: pointer;
+}
+
+.product-price {
+  color: var(--Secondary-800);
+  font-size: 16px;
+  font-style: normal;
+  font-weight: 700;
+  line-height: 26px;
+}
+
+.product-favorite-count {
+  color: var(--Secondary-600);
+  font-size: 12px;
+  font-style: normal;
+  font-weight: 500;
+  line-height: 18px;
+  display: flex;
+  gap: 4px;
+}
+
+.product-heart-img>svg {
+  width: 16px;
+  height: 16px;
+  cursor: pointer;
+}
+
+@media (max-width: 1199px) {
+  .container {
+    width: 100%;
+    padding: 0 24px;
+  }
+
+  .best-product-list {
+    display: flex;
+    justify-content: space-between;
+  }
+
+  .best-product {
+    display: flex;
+    flex-direction: column;
+    gap: 10px;
+    width: 100%;
+    height: 434px;
+  }
+
+  .best-product-img {
+    width: 100%;
+    height: 343px;
+    border-radius: 20.2px;
+  }
+
+  .all-product-list {
+    display: flex;
+    flex-wrap: wrap;
+    column-gap: 16px;
+  }
+
+  .all-product {
+    width: 100%;
+    flex: 1 1 30%;
+  }
+
+  .all-product-info {
+    width: 100%;
+    flex: 1 1 30%;
+  }
+
+  .all-product-img {
+    width: 100%;
+    flex: 1 1 30%;
+    border-radius: 12.9px;
+  }
+}
+
+@media (max-width: 767px) {
+  .container {
+    width: 100%;
+    padding: 0 16px;
+    max-width: 400px;
+  }
+
+  .best-product {
+    display: flex;
+    flex-direction: column;
+    gap: 10px;
+    width: 100%;
+    height: 434px;
+  }
+
+  .best-product-img {
+    width: 100%;
+    height: 343px;
+    border-radius: 20.2px;
+  }
+
+  .all-product-list {
+    margin-top: 71px;
+    display: flex;
+    column-gap: 8px;
+  }
+
+  .all-product {
+    width: 100%;
+    flex: 1 1 45%;
+    height: 264px;
+  }
+
+  .all-product-info {
+    width: 100%;
+    flex: 1 1 45%;
+  }
+
+  .all-product-img {
+    width: 100%;
+    flex: 1 1 45%;
+    height: 167px;
+    border-radius: 9.82px;
+  }
+}
